Add Machine Learning module core

+++ src/Linear/Tensor.cpp
@@ -1,4 +1,4 @@
-#include "Tensor.h"
+#include "MLL/Linear/Tensor.h"
 
 #include <numeric>
 #include <limits>
@@ -38,7 +38,7 @@ ml::Tensor& ml::Tensor::fill(const std::vector<double>& values) {
 }
 
 ml::Tensor& ml::Tensor::randomize() {	// FIXME
-	for (int i = 0; i < dataSize; ++i) { data[i] = 0.0d; }
+	for (int i = 0; i < dataSize; ++i) { data[i] = 0.0; }
     return *this;
 }
 
@@ -147,52 +147,52 @@ ml::Tensor& ml::Tensor::operator/=(const ml::Tensor& other) {
 ml::Tensor ml::Tensor::operator+(double value) const {
 	ml::Tensor result(*this);
 	std::transform(result.data.begin(), result.data.end(), result.data.begin(),
-        [value](int x) { return x + value; });
+        [value](double x) { return x + value; });
 	return result;
 }
 
 ml::Tensor ml::Tensor::operator-(double value) const {
 	ml::Tensor result(*this);
 	std::transform(result.data.begin(), result.data.end(), result.data.begin(),
-        [value](int x) { return x - value; });
+        [value](double x) { return x - value; });
 	return result;
 }
 
 ml::Tensor ml::Tensor::operator*(double value) const {
 	ml::Tensor result(*this);
 	std::transform(result.data.begin(), result.data.end(), result.data.begin(),
-        [value](int x) { return x * value; });
+        [value](double x) { return x * value; });
 	return result;
 }
 
 ml::Tensor ml::Tensor::operator/(double value) const {
 	ml::Tensor result(*this);
 	std::transform(result.data.begin(), result.data.end(), result.data.begin(),
-        [value](int x) { return x / value; });
+        [value](double x) { return x / value; });
 	return result;
 }
 
 ml::Tensor& ml::Tensor::operator+=(double value) {
 	std::transform(data.begin(), data.end(), data.begin(),
-        [value](int x) { return x + value; });
+        [value](double x) { return x + value; });
 	return *this;
 }
 
 ml::Tensor& ml::Tensor::operator-=(double value) {
 	std::transform(data.begin(), data.end(), data.begin(),
-        [value](int x) { return x - value; });
+        [value](double x) { return x - value; });
 	return *this;
 }
 
 ml::Tensor& ml::Tensor::operator*=(double value) {
 	std::transform(data.begin(), data.end(), data.begin(),
-        [value](int x) { return x * value; });
+        [value](double x) { return x * value; });
 	return *this;
 }
 
 ml::Tensor& ml::Tensor::operator/=(double value) {
 	std::transform(data.begin(), data.end(), data.begin(),
-        [value](int x) { return x / value; });
+        [value](double x) { return x / value; });
 	return *this;
 }
